# LibGfx/TIFF: Reduce execution time by 50% to 70%