Before:
```
Benchmark 1: ./Meta/Lagom/BuildLagom/bin/image TIFF/ccitt.tiff --no-output
  Time (mean ± σ):      58.6 ms ±  13.4 ms    [User: 55.9 ms, System: 2.9 ms]
  Range (min … max):    43.2 ms …  78.2 ms    40 runs

Benchmark 1: ./Meta/Lagom/BuildLagom/bin/image TIFF/deflate.tiff --no-output
  Time (mean ± σ):      51.0 ms ±  14.4 ms    [User: 46.4 ms, System: 4.5 ms]
  Range (min … max):    37.8 ms …  75.7 ms    50 runs

Benchmark 1: ../../Meta/Lagom/BuildLagom/bin/TestImageDecoder
  Time (mean ± σ):      1.163 s ±  0.039 s    [User: 1.101 s, System: 0.062 s]
  Range (min … max):    1.127 s …  1.259 s    10 runs
```

After:
```
Benchmark 1: ./Meta/Lagom/BuildLagom/bin/image TIFF/ccitt.tiff --no-output
  Time (mean ± σ):       9.1 ms ±   3.6 ms    [User: 6.2 ms, System: 2.8 ms]
  Range (min … max):     5.5 ms …  17.4 ms    441 runs

Benchmark 1: ./Meta/Lagom/BuildLagom/bin/image TIFF/deflate.tiff --no-output
  Time (mean ± σ):      21.0 ms ±   6.6 ms    [User: 15.6 ms, System: 5.2 ms]
  Range (min … max):     9.3 ms …  35.6 ms    249 runs

Benchmark 1: ../../Meta/Lagom/BuildLagom/bin/TestImageDecoder
  Time (mean ± σ):     625.1 ms ±  18.7 ms    [User: 562.2 ms, System: 62.7 ms]
  Range (min … max):   594.5 ms … 652.3 ms    10 runs
```